pyproject.toml
[tool.poetry]
name = "glam-api"
version = "0.2.0"
description = "NASA Harvest Global Agriculture Monitoring System (GLAM) API"
authors = ["John Keniston <jfkeniston@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10, <4.0"
django = "^4.2"
rio-tiler = "^6.6.1"
django-storages = "^1.14.4"
django-cors-headers = "^4.4.0"
django-rest-framework = "^0.1.0"
django-filter = "^24.2"
pandas = "^2.2.2"
rest-pandas = "^1.1.0"
django-extensions = "^3.2.3"
drf-yasg = "^1.21.7"
django-modeltranslation = "^0.19.5"
django-admin-env-notice = "^1.0"
django-debug-toolbar = "^4.4.6"
django-q = "^1.3.9"
boto3 = "^1.34.148"
matplotlib = "^3.9.1"
shapely = "^2.0.5"
descartes = "^1.1.0"
gunicorn = "^22.0.0"
whitenoise = "^6.7.0"
psycopg2-binary = "^2.9.9"
tqdm = "^4.66.4"
rio-cogeo = "^5.3.3"
rioxarray = "^0.17.0"
ipython = "^8.26.0"
python-semantic-release = "^9.14.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
psycopg2 = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_toml = [
    "pyproject.toml:tool.poetry.version",
]
assets = []
build_command_env = []
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "angular"
logging_use_named_masks = false
major_on_zero = true
allow_zero_version = true
no_git_verify = false
tag_format = "v{version}"

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.changelog]
exclude_commit_patterns = []
mode = "init"
insertion_flag = "<!-- version list -->"
template_dir = "templates"

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
output_format = "md"
mask_initial_release = false

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = false

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
allowed_tags = ["feat", "fix", "perf", "build", "chore", "ci", "docs", "style", "refactor", "test"]
default_bump_level = 0

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false
insecure = false

[tool.semantic_release.remote.token]
env = "GH_TOKEN"

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true

